[Fix] #19 - API 모델 분리
