+++ SOPKATHON-4Team-iOS/SOPKATHON-4Team-iOS/Network/API/AuthAPI.swift
@@ -1,0 +1,12 @@
+//
+//  AuthAPI.swift
+//  SOPKATHON-4Team-iOS
+//
+//  Created by 이명진 on 11/24/24.
+//
+
+import Foundation
+
+struct AuthAPI: Codable {
+    let id: Int
+}
